--- src/components/CircleOfFifths/animations.ts
@@ -1,45 +1,0 @@
-import { AnimationVariants } from '@/types/circleOfFifths';
-
-// キー情報のアニメーション
-export const keyInfoVariants: AnimationVariants = {
-  hidden: {
-    opacity: 0,
-    y: 20,
-    scale: 0.95,
-  },
-  visible: {
-    opacity: 1,
-    y: 0,
-    scale: 1,
-    transition: {
-      duration: 0.3,
-      ease: 'easeOut',
-      staggerChildren: 0.1,
-    },
-  },
-  exit: {
-    opacity: 0,
-    y: -20,
-    scale: 0.95,
-    transition: {
-      duration: 0.2,
-      ease: 'easeIn',
-    },
-  },
-};
-
-// キー情報の子要素のアニメーション
-export const keyInfoItemVariants: AnimationVariants = {
-  hidden: {
-    opacity: 0,
-    x: -20,
-  },
-  visible: {
-    opacity: 1,
-    x: 0,
-    transition: {
-      duration: 0.3,
-      ease: 'easeOut',
-    },
-  },
-};
