+++ src/features/circle-of-fifths/__stories__/CircleOfFifths.stories.tsx
@@ -1,0 +1,18 @@
+import type { Meta, StoryObj } from '@storybook/react';
+import { CircleOfFifths } from '../components/CircleOfFifths';
+
+const meta: Meta<typeof CircleOfFifths> = {
+  title: 'Components/CircleOfFifths',
+  component: CircleOfFifths,
+  parameters: {
+    layout: 'centered',
+  },
+  tags: ['autodocs'],
+};
+
+export default meta;
+type Story = StoryObj<typeof meta>;
+
+export const Default: Story = {
+  args: {},
+};
