+++ src/features/circle-of-fifths/index.ts
@@ -1,0 +1,1 @@
+export * from './components/CircleOfFifths';
