+++ src/features/global-header/components/Logo.tsx
