+++ src/features/global-header/hooks/useMobileMenu.ts
