+++ src/features/global-header/types.ts
